api-service/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pymongo>=4.6.0
motor>=3.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.0
Pillow>=10.0.0

# Development & Code Quality
ruff>=0.1.0
flake8>=6.1.0
black>=23.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
